[TEST] Move getWorking*Params to separate ParamsFixture
