+++ tests/Request/InitTest.php
@@ -2,8 +2,9 @@
 
 namespace LearnositySdk\Request;
 
-use LearnositySdk\AbstractTestCase;
+use LearnositySdk\AbstractTestCase; // XXX: should be in a Test namespace
 use LearnositySdk\Exceptions\ValidationException;
+use LearnositySdk\Fixtures\ParamsFixture; // XXX: should be in a Test namespace
 use LearnositySdk\Services\SignatureFactory;
 
 class InitTest extends AbstractTestCase
@@ -120,35 +121,35 @@ class InitTest extends AbstractTestCase
 
     public function testNullRequestPacketGeneratesValidInit()
     {
-        list($service, $security, $secret) = static::getWorkingDataApiParams();
+        list($service, $security, $secret) = ParamsFixture::getWorkingDataApiParams();
         $initObject = new Init($service, $security, $secret, null, null);
         $this->assertInstanceOf(Init::class, $initObject);
     }
 
     public function testEmptyArrayRequestPacketGeneratesValidInit()
     {
-        list($service, $security, $secret) = static::getWorkingDataApiParams();
+        list($service, $security, $secret) = ParamsFixture::getWorkingDataApiParams();
         $initObject = new Init($service, $security, $secret, [], null);
         $this->assertInstanceOf(Init::class, $initObject);
     }
 
     public function testEmptyStringGeneratesValidInit()
     {
-        list($service, $security, $secret) = static::getWorkingDataApiParams();
+        list($service, $security, $secret) = ParamsFixture::getWorkingDataApiParams();
         $initObject = new Init($service, $security, $secret, "", null);
         $this->assertInstanceOf(Init::class, $initObject);
     }
 
     public function testNullRequestPacketAndActionGeneratesValidInit()
     {
-        list($service, $security, $secret) = static::getWorkingDataApiParams();
+        list($service, $security, $secret) = ParamsFixture::getWorkingDataApiParams();
         $initObject = new Init($service, $security, $secret, null, null);
         $this->assertInstanceOf(Init::class, $initObject);
     }
 
     public function testMetaWithTelemetryOnlyAddsSdkProp()
     {
-        list($service, $security, $secret, $request, $action) = static::getWorkingQuestionsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingQuestionsApiParams();
 
         // This test verifies the correctness of the added telemetry data
         Init::enableTelemetry();
@@ -165,10 +166,10 @@ class InitTest extends AbstractTestCase
 
     public function testRequestWithTelemetryPreservesOtherMetaProps()
     {
-        list($service, $security, $secret, $request, $action) = static::getWorkingQuestionsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingQuestionsApiParams();
 
         // add meta field to the $request
-        $request['meta'] = $this->getMetaField();
+        $request['meta'] = ParamsFixture::getMetaField();
 
         // generate a new $initObject using the updated $request
         $initObject = new Init($service, $security, $secret, json_encode($request), $action);
@@ -191,7 +192,7 @@ class InitTest extends AbstractTestCase
     public function testRequestWithoutTelemetryPreservesEmptyMeta()
     {
         Init::disableTelemetry();
-        list($service, $security, $secret, $request, $action) = static::getWorkingQuestionsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingQuestionsApiParams();
 
         $initObject = new Init($service, $security, $secret, json_encode($request), $action);
         $generatedObject = json_decode($initObject->generate());
@@ -204,10 +205,10 @@ class InitTest extends AbstractTestCase
     public function testRequestWithoutTelemetryPreservesFilledMeta()
     {
         Init::disableTelemetry();
-        list($service, $security, $secret, $request, $action) = static::getWorkingQuestionsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingQuestionsApiParams();
 
         // add meta field to the $request
-        $request['meta'] = $this->getMetaField();
+        $request['meta'] = ParamsFixture::getMetaField();
 
         $initObject = new Init($service, $security, $secret, json_encode($request), $action);
         $generatedObject = json_decode($initObject->generate());
@@ -227,445 +228,12 @@ class InitTest extends AbstractTestCase
     }
 
     /*
-     * Pseudo fixtures for parameters
-     */
-
-    /**
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingAssessApiParams(bool $assoc = false): array
-    {
-        $service = 'assess';
-        $security = static::getSecurity();
-        // Needed to initialise Questions API
-        $security['user_id'] = '$ANONYMIZED_USER_ID';
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            "items" => [
-                [
-                    "content" => '<span class="learnosity-response question-demoscience1234"></span>',
-                    "response_ids" => [
-                        "demoscience1234"
-                    ],
-                    "workflow" => "",
-                    "reference" => "question-demoscience1"
-                ],
-                [
-                    "content" => '<span class="learnosity-response question-demoscience5678"></span>',
-                    "response_ids" => [
-                        "demoscience5678"
-                    ],
-                    "workflow" => "",
-                    "reference" => "question-demoscience2"
-                ]
-            ],
-            "ui_style" => "horizontal",
-            "name" => "Demo (2 questions)",
-            "state" => "initial",
-            "metadata" => [],
-            "navigation" => [
-                "show_next" => true,
-                "toc" => true,
-                "show_submit" => true,
-                "show_save" => false,
-                "show_prev" => true,
-                "show_title" => true,
-                "show_intro" => true,
-            ],
-            "time" => [
-                "max_time" => 600,
-                "limit_type" => "soft",
-                "show_pause" => true,
-                "warning_time" => 60,
-                "show_time" => true
-            ],
-            "configuration" => [
-                "onsubmit_redirect_url" => "/assessment/",
-                "onsave_redirect_url" => "/assessment/",
-                "idle_timeout" => true,
-                "questionsApiVersion" => "v2"
-            ],
-            "questionsApiActivity" => [
-                "user_id" => '$ANONYMIZED_USER_ID',
-                "type" => "submit_practice",
-                "state" => "initial",
-                "id" => "assessdemo",
-                "name" => "Assess API - Demo",
-                "questions" => [
-                    [
-                        "response_id" => "demoscience1234",
-                        "type" => "sortlist",
-                        "description" => "In this question, the student needs to sort the events, chronologically earliest to latest.",
-                        "list" => ["Russian Revolution", "Discovery of the Americas", "Storming of the Bastille", "Battle of Plataea", "Founding of Rome", "First Crusade"],
-                        "instant_feedback" => true,
-                        "feedback_attempts" => 2,
-                        "validation" => [
-                            "valid_response" => [4, 3, 5, 1, 2, 0],
-                            "valid_score" => 1,
-                            "partial_scoring" => true,
-                            "penalty_score" => -1
-                        ]
-                    ],
-                    [
-                        "response_id" => "demoscience5678",
-                        "type" => "highlight",
-                        "description" => "The student needs to mark one of the flowers anthers in the image.",
-                        "img_src" => "http://www.learnosity.com/static/img/flower.jpg",
-                        "line_color" => "rgb(255, 20, 0)",
-                        "line_width" => "4"
-                    ]
-                ]
-            ],
-            "type" => "activity"
-        ];
-        $action = null;
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action,
-            ];
-        }
-    }
-
-    /**
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingAuthorApiParams(bool $assoc = false): array
-    {
-        $service = 'author';
-        $security = static::getSecurity();
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            "mode" => "item_list",
-            "config" => [
-                "item_list" => [
-                    "item" => [
-                        "status" => true
-                    ]
-                ]
-            ],
-            "user" => [
-                "id" => "walterwhite",
-                "firstname" => "walter",
-                "lastname" => "white"
-            ]
-        ];
-        $action = null;
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action,
-            ];
-        }
-    }
-
-    /**
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingAuthorAideApiParams(bool $assoc = false): array
-    {
-        $service = 'authoraide';
-        $security = static::getSecurity();
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            "config" => [
-                "test-attribute" => "test"
-            ],
-            "user" => [
-                "id" => "walterwhite",
-                "firstname" => "walter",
-                "lastname" => "white"
-            ]
-        ];
-        $action = null;
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action,
-            ];
-        }
-    }
-
-    /**
-     * WARNING: RemoteTest is also using this params
-     *
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingDataApiParams(bool $assoc = false): array
-    {
-        $service = 'data';
-        $security = static::getSecurity();
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            'limit' => 100,
-        ];
-        $action = 'get';
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action
-            ];
-        }
-    }
-
-    /**
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingEventsApiParams(bool $assoc = false): array
-    {
-        $service = 'events';
-        $security = static::getSecurity();
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            'users' => [
-                '$ANONYMIZED_USER_ID_1' => '$ANONYMIZED_USER_ID_1_NAME',
-                '$ANONYMIZED_USER_ID_2' => '$ANONYMIZED_USER_ID_2_NAME',
-                '$ANONYMIZED_USER_ID_3' => '$ANONYMIZED_USER_ID_3_NAME',
-                '$ANONYMIZED_USER_ID_4' => '$ANONYMIZED_USER_ID_4_NAME'
-            ]
-        ];
-        $action = null;
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action,
-            ];
-        }
-    }
-
-    /**
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingItemsApiParams(bool $assoc = false): array
-    {
-        $service = 'items';
-        $security = static::getSecurity();
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            'user_id' => '$ANONYMIZED_USER_ID',
-            'rendering_type' => 'assess',
-            'name' => 'Items API demo - assess activity demo',
-            'state' => 'initial',
-            'activity_id' => 'items_assess_demo',
-            'session_id' => 'demo_session_uuid',
-            'type' => 'submit_practice',
-            'config' => [
-                'configuration' => [
-                    'responsive_regions' => true
-                ],
-                'navigation' => [
-                    'scrolling_indicator' => true
-                ],
-                'regions' => 'main',
-                'time' => [
-                    'show_pause' => true,
-                    'max_time' => 300
-                ],
-                'title' => 'ItemsAPI Assess Isolation Demo',
-                'subtitle' => 'Testing Subtitle Text'
-            ],
-            'items' => [
-                'Demo3'
-            ]
-        ];
-        $action = null;
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action,
-            ];
-        }
-    }
-
-    /**
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingQuestionsApiParams(bool $assoc = false): array
-    {
-        $service = 'questions';
-        $security = static::getSecurity();
-        $security['user_id'] = '$ANONYMIZED_USER_ID';
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            'type'      => 'local_practice',
-            'state'     => 'initial',
-            'questions' => [
-                [
-                    'response_id'        => '60005',
-                    'type'               => 'association',
-                    'stimulus'           => 'Match the cities to the parent nation.',
-                    'stimulus_list'      => ['London', 'Dublin', 'Paris', 'Sydney'],
-                    'possible_responses' => ['Australia', 'France', 'Ireland', 'England'],
-                    'validation' => [
-                        'valid_responses' => [
-                            ['England'], ['Ireland'], ['France'], ['Australia']
-                        ]
-                    ]
-                ]
-            ]
-        ];
-        $action = null;
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action,
-            ];
-        }
-    }
-
-    /**
-     * @param  bool $assoc If true, associative array will be returned
-     * @return array
-     */
-    public static function getWorkingReportsApiParams(bool $assoc = false): array
-    {
-        $service = 'reports';
-        $security = static::getSecurity();
-        $secret = static::TEST_CONSUMER_SECRET;
-        $request = [
-            "reports" => [
-                [
-                    "id" => "report-1",
-                    "type" => "sessions-summary",
-                    "user_id" => '$ANONYMIZED_USER_ID',
-                    "session_ids" => [
-                        "AC023456-2C73-44DC-82DA28894FCBC3BF"
-                    ]
-                ]
-            ]
-        ];
-        $action = null;
-
-        if ($assoc) {
-            return [
-                'service' => $service,
-                'security' => $security,
-                'secret' => $secret,
-                'request' => $request,
-                'action' => $action,
-            ];
-        } else {
-            return [
-                $service,
-                $security,
-                $secret,
-                $request,
-                $action,
-            ];
-        }
-    }
-
-    /**
-     *
-     * @return array
-     */
-    public function getMetaField(): array
-    {
-        return [
-            "test_key_string" => "test-string",
-            "test_key_integer" => 12345,
-            "test_key_boolean" => true,
-        ];
-    }
-
-    /*
      * Data providers
      */
 
     public function constructorProvider(): array
     {
-        list($service, $security, $secret, $request, $action) = static::getWorkingDataApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingDataApiParams();
 
         $wrongSecurity = $security;
         $wrongSecurity['wrongParam'] = '';
@@ -697,7 +265,7 @@ class InitTest extends AbstractTestCase
         $testCases = [];
 
         /* Author */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAuthorApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAuthorApiParams();
         $authorApi = [
             '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","signature":"$02$ca2769c4be77037cf22e0f7a2291fe48c470ac6db2f45520a259907370eff861"},"request":{"mode":"item_list","config":{"item_list":{"item":{"status":true}}},"user":{"id":"walterwhite","firstname":"walter","lastname":"white"}}}',
             $service, $security, $secret, $request, $action,
@@ -705,7 +273,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-author'] = $authorApi;
 
         /* Author Aide */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAuthorAideApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAuthorAideApiParams();
         $authorAideApi = [
             '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","signature":"$02$f2ce1da2fdead193d53ab954b8a3660548ed9b0e3ce60599d751130deba7a138"},"request":{"config":{"test-attribute":"test"},"user":{"id":"walterwhite","firstname":"walter","lastname":"white"}}}',
             $service, $security, $secret, $request, $action,
@@ -713,7 +281,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-authoraide'] = $authorAideApi;
 
         /* Assess */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAssessApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAssessApiParams();
         $assessApi = [
             '{"items":[{"content":"<span class=\"learnosity-response question-demoscience1234\"></span>","response_ids":["demoscience1234"],"workflow":"","reference":"question-demoscience1"},{"content":"<span class=\"learnosity-response question-demoscience5678\"></span>","response_ids":["demoscience5678"],"workflow":"","reference":"question-demoscience2"}],"ui_style":"horizontal","name":"Demo (2 questions)","state":"initial","metadata":[],"navigation":{"show_next":true,"toc":true,"show_submit":true,"show_save":false,"show_prev":true,"show_title":true,"show_intro":true},"time":{"max_time":600,"limit_type":"soft","show_pause":true,"warning_time":60,"show_time":true},"configuration":{"onsubmit_redirect_url":"/assessment/","onsave_redirect_url":"/assessment/","idle_timeout":true,"questionsApiVersion":"v2"},"questionsApiActivity":{"user_id":"$ANONYMIZED_USER_ID","type":"submit_practice","state":"initial","id":"assessdemo","name":"Assess API - Demo","questions":[{"response_id":"demoscience1234","type":"sortlist","description":"In this question, the student needs to sort the events, chronologically earliest to latest.","list":["Russian Revolution","Discovery of the Americas","Storming of the Bastille","Battle of Plataea","Founding of Rome","First Crusade"],"instant_feedback":true,"feedback_attempts":2,"validation":{"valid_response":[4,3,5,1,2,0],"valid_score":1,"partial_scoring":true,"penalty_score":-1}},{"response_id":"demoscience5678","type":"highlight","description":"The student needs to mark one of the flowers anthers in the image.","img_src":"http://www.learnosity.com/static/img/flower.jpg","line_color":"rgb(255, 20, 0)","line_width":"4"}],"consumer_key":"yis0TYCu7U9V4o7M","timestamp":"20140626-0528","signature":"$02$8de51b7601f606a7f32665541026580d09616028dde9a929ce81cf2e88f56eb8"},"type":"activity"}',
             $service, $security, $secret, $request, $action,
@@ -721,7 +289,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-assess'] = $assessApi;
 
         /* Data */
-        list($service, $security, $secret, $request, $action) = static::getWorkingDataApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingDataApiParams();
         $security['timestamp'] = '20140626-0528';
         $dataApiGet = [
             [
@@ -744,7 +312,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-data_post'] = $dataApiPost;
 
         /* Events */
-        list($service, $security, $secret, $request, $action) = static::getWorkingEventsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingEventsApiParams();
         $eventsApiExpected = '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","signature":"$02$5c3160dbb9ab4d01774b5c2fc3b01a35ce4f9709c84571c27dfe333d1ca9d349"},"config":{"users":{"$ANONYMIZED_USER_ID_1":"64ccf06154cf4133624372459ebcccb8b2f8bd7458a73df681acef4e742e175c","$ANONYMIZED_USER_ID_2":"7fa4d6ef8926add8b6411123fce916367250a6a99f50ab8ec39c99d768377adb","$ANONYMIZED_USER_ID_3":"3d5b26843da9192319036b67f8c5cc26e1e1763811270ba164665d0027296952","$ANONYMIZED_USER_ID_4":"3b6ac78f60f3e3eb7a85cec8b48bdca0f590f959e0a87a9c4222898678bd50c8"}}}';
         $eventsApi = [
             $eventsApiExpected,
@@ -753,7 +321,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-events'] = $eventsApi;
 
         /* Items */
-        list($service, $security, $secret, $request, $action) = static::getWorkingItemsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingItemsApiParams();
         $itemsApi = [
             '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","user_id":"$ANONYMIZED_USER_ID","signature":"$02$36c439e7d18f2347ce08ca4b8d4803a22325d54352650b19b6f4aaa521b613d9"},"request":{"user_id":"$ANONYMIZED_USER_ID","rendering_type":"assess","name":"Items API demo - assess activity demo","state":"initial","activity_id":"items_assess_demo","session_id":"demo_session_uuid","type":"submit_practice","config":{"configuration":{"responsive_regions":true},"navigation":{"scrolling_indicator":true},"regions":"main","time":{"show_pause":true,"max_time":300},"title":"ItemsAPI Assess Isolation Demo","subtitle":"Testing Subtitle Text"},"items":["Demo3"]}}',
             $service, $security, $secret, $request, $action,
@@ -761,7 +329,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-items'] = $itemsApi;
 
         /* Questions */
-        list($service, $security, $secret, $request, $action) = static::getWorkingQuestionsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingQuestionsApiParams();
         $questionsApi = [
             '{"consumer_key":"yis0TYCu7U9V4o7M","timestamp":"20140626-0528","user_id":"$ANONYMIZED_USER_ID","signature":"$02$8de51b7601f606a7f32665541026580d09616028dde9a929ce81cf2e88f56eb8","type":"local_practice","state":"initial","questions":[{"response_id":"60005","type":"association","stimulus":"Match the cities to the parent nation.","stimulus_list":["London","Dublin","Paris","Sydney"],"possible_responses":["Australia","France","Ireland","England"],"validation":{"valid_responses":[["England"],["Ireland"],["France"],["Australia"]]}}]}',
             $service, $security, $secret, $request, $action,
@@ -769,7 +337,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-questions'] = $questionsApi;
 
         /* Reports */
-        list($service, $security, $secret, $request, $action) = static::getWorkingReportsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingReportsApiParams();
         $reportsApi = [
             '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","signature":"$02$8e0069e7aa8058b47509f35be236c53fa1a878c64b12589fd42f48b568f6ac84"},"request":{"reports":[{"id":"report-1","type":"sessions-summary","user_id":"$ANONYMIZED_USER_ID","session_ids":["AC023456-2C73-44DC-82DA28894FCBC3BF"]}]}}',
             $service, $security, $secret, $request, $action,
@@ -777,14 +345,14 @@ class InitTest extends AbstractTestCase
         $testCases['api-reports'] = $reportsApi;
 
         /* Passing request as string */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAuthorApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAuthorApiParams();
         $authorApiAsString = [
             '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","signature":"$02$ca2769c4be77037cf22e0f7a2291fe48c470ac6db2f45520a259907370eff861"},"request":"{\"mode\":\"item_list\",\"config\":{\"item_list\":{\"item\":{\"status\":true}}},\"user\":{\"id\":\"walterwhite\",\"firstname\":\"walter\",\"lastname\":\"white\"}}"}',
             $service, $security, $secret, json_encode($request), $action,
         ];
         $testCases['api-author_string'] = $authorApiAsString;
 
-        list($service, $security, $secret, $request, $action) = static::getWorkingAuthorAideApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAuthorAideApiParams();
         $authorAideApiAsString = [
             '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","signature":"$02$f2ce1da2fdead193d53ab954b8a3660548ed9b0e3ce60599d751130deba7a138"},"request":"{\"config\":{\"test-attribute\":\"test\"},\"user\":{\"id\":\"walterwhite\",\"firstname\":\"walter\",\"lastname\":\"white\"}}"}',
             $service, $security, $secret, json_encode($request), $action,
@@ -792,7 +360,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-authoraide_string'] = $authorAideApiAsString;
 
         /* Items */
-        list($service, $security, $secret, $request, $action) = static::getWorkingItemsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingItemsApiParams();
         $itemsApiAsString = [
             '{"security":{"consumer_key":"yis0TYCu7U9V4o7M","domain":"localhost","timestamp":"20140626-0528","user_id":"$ANONYMIZED_USER_ID","signature":"$02$36c439e7d18f2347ce08ca4b8d4803a22325d54352650b19b6f4aaa521b613d9"},"request":"{\"user_id\":\"$ANONYMIZED_USER_ID\",\"rendering_type\":\"assess\",\"name\":\"Items API demo - assess activity demo\",\"state\":\"initial\",\"activity_id\":\"items_assess_demo\",\"session_id\":\"demo_session_uuid\",\"type\":\"submit_practice\",\"config\":{\"configuration\":{\"responsive_regions\":true},\"navigation\":{\"scrolling_indicator\":true},\"regions\":\"main\",\"time\":{\"show_pause\":true,\"max_time\":300},\"title\":\"ItemsAPI Assess Isolation Demo\",\"subtitle\":\"Testing Subtitle Text\"},\"items\":[\"Demo3\"]}"}',
             $service, $security, $secret, json_encode($request), $action,
@@ -817,7 +385,7 @@ class InitTest extends AbstractTestCase
         $testCases = [];
 
         /* Author */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAuthorApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAuthorApiParams();
         $authorApi = [
             '$02$ca2769c4be77037cf22e0f7a2291fe48c470ac6db2f45520a259907370eff861',
             $service, $security, $secret, $request, $action,
@@ -825,7 +393,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-author'] = $authorApi;
 
         /* Assess */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAssessApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAssessApiParams();
         $assessApi = [
             '$02$8de51b7601f606a7f32665541026580d09616028dde9a929ce81cf2e88f56eb8',
             $service, $security, $secret, $request, $action,
@@ -833,7 +401,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-assess'] = $assessApi;
 
         /* Data */
-        list($service, $security, $secret, $request, $action) = static::getWorkingDataApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingDataApiParams();
         $securityExpires = $security;
         $securityExpires['expires'] = '20160621-1716';
 
@@ -856,7 +424,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-data_expire'] = $dataApiExpire;
 
         /* Events */
-        list($service, $security, $secret, $request, $action) = static::getWorkingEventsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingEventsApiParams();
         $eventsApi = [
             '$02$5c3160dbb9ab4d01774b5c2fc3b01a35ce4f9709c84571c27dfe333d1ca9d349',
             $service, $security, $secret, $request, $action,
@@ -864,7 +432,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-events'] = $eventsApi;
 
         /* Items */
-        list($service, $security, $secret, $request, $action) = static::getWorkingItemsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingItemsApiParams();
         $itemsApi = [
             '$02$36c439e7d18f2347ce08ca4b8d4803a22325d54352650b19b6f4aaa521b613d9',
             $service, $security, $secret, $request, $action,
@@ -872,7 +440,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-items'] = $itemsApi;
 
         /* Questions */
-        list($service, $security, $secret, $request, $action) = static::getWorkingQuestionsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingQuestionsApiParams();
         $questionsApi = [
             '$02$8de51b7601f606a7f32665541026580d09616028dde9a929ce81cf2e88f56eb8',
             $service, $security, $secret, $request, $action,
@@ -880,7 +448,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-questions'] = $questionsApi;
 
         /* Reports */
-        list($service, $security, $secret, $request, $action) = static::getWorkingReportsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingReportsApiParams();
         $reportsApi = [
             '$02$8e0069e7aa8058b47509f35be236c53fa1a878c64b12589fd42f48b568f6ac84',
             $service, $security, $secret, $request, $action,
@@ -903,7 +471,7 @@ class InitTest extends AbstractTestCase
         $testCases = [];
 
         /* Author */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAuthorApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAuthorApiParams();
         $authorApi = [
             'request.meta',
             $service, $security, $secret, $request, $action,
@@ -911,7 +479,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-author'] = $authorApi;
 
         /* Author Aide */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAuthorAideApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAuthorAideApiParams();
         $authorAideApi = [
             'request.meta',
             $service, $security, $secret, $request, $action,
@@ -919,7 +487,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-authoraide'] = $authorAideApi;
 
         /* Assess */
-        list($service, $security, $secret, $request, $action) = static::getWorkingAssessApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingAssessApiParams();
         $assessApi = [
             'meta',
             $service, $security, $secret, $request, $action,
@@ -927,7 +495,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-assess'] = $assessApi;
 
         /* Data */
-        list($service, $security, $secret, $request, $action) = static::getWorkingDataApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingDataApiParams();
         $dataApi = [
             'request.meta',
             $service, $security, $secret, $request, $action,
@@ -935,7 +503,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-data'] = $dataApi;
 
         /* Events */
-        list($service, $security, $secret, $request, $action) = static::getWorkingEventsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingEventsApiParams();
         $eventsApi = [
             'config.meta',
             $service, $security, $secret, $request, $action,
@@ -943,7 +511,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-events'] = $eventsApi;
 
         /* Items */
-        list($service, $security, $secret, $request, $action) = static::getWorkingItemsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingItemsApiParams();
         $itemsApi = [
             'request.meta',
             $service, $security, $secret, $request, $action,
@@ -951,7 +519,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-items'] = $itemsApi;
 
         /* Questions */
-        list($service, $security, $secret, $request, $action) = static::getWorkingQuestionsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingQuestionsApiParams();
         $questionsApi = [
             'meta',
             $service, $security, $secret, $request, $action,
@@ -959,7 +527,7 @@ class InitTest extends AbstractTestCase
         $testCases['api-questions'] = $questionsApi;
 
         /* Reports */
-        list($service, $security, $secret, $request, $action) = static::getWorkingReportsApiParams();
+        list($service, $security, $secret, $request, $action) = ParamsFixture::getWorkingReportsApiParams();
         $reportsApi = [
             'request.meta',
             $service, $security, $secret, $request, $action,
